requirements.txt
datasets==2.14.5
pyarrow>=12.0.0,<14.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain>=0.1.17,<0.2.0
langchain-mistralai>=0.1.0
mistralai>=0.1.3
streamlit==1.28.1
python-dotenv==1.0.0
pytest==7.4.3
flake8==6.1.0
black==23.11.0
isort==5.12.0
rouge-score==0.1.2
tqdm==4.66.1
huggingface-hub>=0.19.0,<0.24.0